+++ src/compas_timber/fabrication/joint_factories/l_butt_factory.py
@@ -1,6 +1,7 @@
 from compas_timber.connections import LButtJoint
 from compas_timber.fabrication import BTLx
 from compas_timber.fabrication import BTLxJackCut
+from compas_timber.fabrication import BTLxLap
 
 
 class LButtFactory(object):
@@ -28,10 +29,31 @@ class LButtFactory(object):
 
         """
 
-        main_part = parts[str(joint.main_beam.key)]
-        cross_part = parts[str(joint.cross_beam.key)]
-        main_part.processings.append(BTLxJackCut.create_process(main_part, joint.cutting_plane_main, "L-Butt Joint"))
-        cross_part.processings.append(BTLxJackCut.create_process(cross_part, joint.cutting_plane_cross, "L-Butt Joint"))
+        main_part = parts[str(joint.main_beam.guid)]
+        cross_part = parts[str(joint.cross_beam.guid)]
+        ref_plane_cross = joint.get_main_cutting_plane()[1]
+        main_part.processings.append(
+            BTLxJackCut.create_process(main_part, joint.get_main_cutting_plane()[0], "L-Butt Joint")
+        )
+        cross_part.processings.append(
+            BTLxJackCut.create_process(cross_part, joint.get_cross_cutting_plane(), "L-Butt Joint")
+        )
+        if joint.mill_depth > 0:
+            if joint.ends[str(cross_part.part_guid)] == "start":
+                # TODO: aren't these two the same?
+                joint.btlx_params_cross["machining_limits"] = {
+                    "FaceLimitedStart": "no",
+                    "FaceLimitedFront": "no",
+                    "FaceLimitedBack": "no",
+                }
+            else:
+                joint.btlx_params_cross["machining_limits"] = {
+                    "FaceLimitedEnd": "no",
+                    "FaceLimitedFront": "no",
+                    "FaceLimitedBack": "no",
+                }
+            joint.btlx_params_cross["reference_plane_id"] = cross_part.ref_side_from_face(ref_plane_cross)
+            cross_part.processings.append(BTLxLap.create_process(joint.btlx_params_cross, "L-Butt Joint"))
 
 
 BTLx.register_joint(LButtJoint, LButtFactory)